moved AffinityPropagation to a clustering package.
