+++ src/main/java/ml/grafos/okapi/clustering/ap/AffinityPropagation.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ml.grafos.okapi.graphs;
+package ml.grafos.okapi.clustering.ap;
 
 import com.google.common.collect.ComparisonChain;
 import es.csic.iiia.bms.CommunicationAdapter;
