+++ src/test/java/ml/grafos/okapi/clustering/ap/AffinityPropagationTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ml.grafos.okapi.graphs;
+package ml.grafos.okapi.clustering.ap;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Ordering;
